Update index.html
Gemini 2.5 Pro prompt:

this feels like a pretty solid game. Let's do a few adjustments to help with progression through the game.



- Speed isn't actually that useful. I think engine upgrades should include some sort of cooling boost or armor boost instead of speed. Maybe they are more of a "core" upgrade not only engine. NOTE: There are a few players already playing the game so we should try to be non-breaking with these changes.



- The drop rates on each level are very discouraging as is. I suggest that each wave cleared yields a drop. At the end of the level when it shows the summary screen there should be a slot per wave plus a level clear slot. Each one of these slots for the 10 levels should have their own probability of dropping rare loot, but each should at least drop a common. Clearing the level should have the highest percentage chance of dropping rarer loot.



Also, the coins and materials earned per level are too low when you successfully clear it. Instead of a "win" and "defeat" amount of coins and materials, let's make coins and materials come out of asteroids when the small ones are finally defeated. When the ship runs into an asteroid there shouldn't be any coins or materials earned. Add a UI element when a small asteroid is shot that puts a brief "+" in yellow for coins and/or green for materials based on what was earned. This will give it more of an arcade feel. Make this amount random per asteroid.


- Also my inventory is starting to get unweildy with many duplicates of common items. Can you stack multiples of the same item+level?