Update index.html
Gemini 2.5 Pro Prompt:

the drop rates still seem very low for rarer items. I'm not sure if there is a way around this. Maybe if you get 10 of a common item you can refine it with enough coins and materials into an uncommon item? And the same with 10 uncommons to a rare etc? As an addition to the recipe mechanic?



Also, let's put recipes in the wave clearing loot crates rather than randomly as the level plays live/asteroids are cleared. This will make more sense to the user